fixed growing with video and better track management
